adding cosmwasm-vm kernel for notebook
